+++ packages/logger/jest.config.js
@@ -14,4 +14,16 @@ module.exports = {
     '!src/__mocks__/**',
   ],
   clearMocks: true,
+  transform: {
+    '^.+\\.tsx?$': [
+      'ts-jest',
+      {
+        useESM: false,
+        tsconfig: {
+          esModuleInterop: true,
+          allowSyntheticDefaultImports: true,
+        },
+      },
+    ],
+  },
 };